image-metadata-pipeline/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.34.0
azure-eventhub==5.15.0
azure-functions==1.23.0
azure-functions-durable==1.3.2
azure-identity==1.23.0
azure-iot-device==2.14.0
azure-search-documents==11.5.3
azure-storage-blob==12.25.1
backoff==2.2.1
beautifulsoup4==4.13.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.3
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
emoji==2.14.1
fastapi==0.115.12
fastapi-cli==0.0.7
filetype==1.2.0
Flask==2.3.2
Flask-Cors==5.0.0
frozenlist==1.7.0
furl==2.1.4
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.5
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.6
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.5.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.2
oauthlib==3.3.0
olefile==0.47
openai==1.96.1
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orderedmultidict==1.0.1
orjson==3.11.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
pypdf==5.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests-unixsocket2==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.18.7
unstructured-client==0.38.1
urllib3==2.3.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
